Add support for multi partitioned reads in schema and label scan store lucene indexes.
Introduce readers for cases when index contains more then one partition.
Readers will go over all available partitions and combine results
so index clients should not observe any difference compare to single partitioned index.
There is no guarantee in which order client will receive results from different partitions.
Add possibility to perform multi partitioned sampling.
Unit and integration test for new readers provided.

Currently for all partitioned tasks used common forkJoinPool, that can change after performance impact testing/measurements

+++ community/consistency-check-legacy/src/test/java/org/neo4j/legacy/consistency/checking/full/NodeCorrectlyIndexedCheckTest.java
@@ -253,7 +253,7 @@ public class NodeCorrectlyIndexedCheckTest
                 }
 
                 @Override
-                public int countIndexedNodes( long nodeId, Object propertyValue )
+                public long countIndexedNodes( long nodeId, Object propertyValue )
                 {
                     long[] candidates = entries.get( propertyValue );
                     if ( candidates == null )
